+++ sources/relay.h
@@ -9,8 +9,8 @@
 
 typedef struct od_relay od_relay_t;
 
-typedef od_status_t (*od_relay_on_packet_t)(od_relay_t*, char *data, int size);
-typedef void        (*od_relay_on_read_t)(od_relay_t*, int size);
+typedef od_status_t (*od_relay_on_packet_t)(od_relay_t*, char *data, size_t size);
+typedef void        (*od_relay_on_read_t)(od_relay_t*, size_t size);
 
 struct od_relay
 {
@@ -176,7 +176,7 @@ od_relay_on_packet_msg(od_relay_t *relay, machine_msg_t *msg)
 }
 
 static inline od_status_t
-od_relay_on_packet(od_relay_t *relay, char *data, int size)
+od_relay_on_packet(od_relay_t *relay, char *data, size_t size)
 {
 	int rc;
 	od_status_t status;
@@ -199,7 +199,7 @@ od_relay_on_packet(od_relay_t *relay, char *data, int size)
 }
 
 __attribute__((hot)) static inline od_status_t
-od_relay_process(od_relay_t *relay, int *progress, char *data, int size)
+od_relay_process(od_relay_t *relay, int *progress, char *data, size_t size)
 {
 	*progress = 0;
